requirements.txt
gunicorn==21.2.0
Flask==2.3.3
Werkzeug==2.3.7
PyMySQL==1.1.0
python-dotenv==1.0.0
cryptography==41.0.7